requirements.txt
lxml
requests
bs4
peewee
telepot
pytz
schedule
python-dateutil
retrying
emoji
